chore(resolver): move `Resolver` to `resolve.rs`
